+++ assets/Magic Market/Tilesets/Objects.tsx
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="Objects" tilewidth="64" tileheight="48" tilecount="2" columns="0">
+ <grid orientation="orthogonal" width="1" height="1"/>
+ <tile id="0">
+  <image source="../Art/Trailer A - 64x32.png" width="64" height="32"/>
+ </tile>
+ <tile id="1">
+  <image source="../Art/Tree A - 48x48.png" width="48" height="48"/>
+ </tile>
+</tileset>