+++ assets/Magic Market/Tilesets/Tileset props - 16x16.tsx
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="Tileset props - 16x16" tilewidth="16" tileheight="16" tilecount="49" columns="7">
+ <image source="../Art/Tileset props - 16x16.png" width="112" height="112"/>
+ <tile id="3">
+  <animation>
+   <frame tileid="0" duration="200"/>
+   <frame tileid="1" duration="200"/>
+   <frame tileid="2" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="11">
+  <animation>
+   <frame tileid="7" duration="200"/>
+   <frame tileid="8" duration="200"/>
+   <frame tileid="9" duration="200"/>
+   <frame tileid="10" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="18">
+  <animation>
+   <frame tileid="14" duration="200"/>
+   <frame tileid="15" duration="200"/>
+   <frame tileid="16" duration="200"/>
+   <frame tileid="17" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="23">
+  <animation>
+   <frame tileid="21" duration="200"/>
+   <frame tileid="22" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="26">
+  <animation>
+   <frame tileid="24" duration="200"/>
+   <frame tileid="25" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="31">
+  <animation>
+   <frame tileid="28" duration="200"/>
+   <frame tileid="29" duration="200"/>
+   <frame tileid="30" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="39">
+  <animation>
+   <frame tileid="35" duration="200"/>
+   <frame tileid="36" duration="200"/>
+   <frame tileid="37" duration="200"/>
+   <frame tileid="38" duration="200"/>
+  </animation>
+ </tile>
+</tileset>