+++ assets/Magic Market/Tilesets/Tileset props - 16x32.tsx
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="Tileset props - 16x32" tilewidth="16" tileheight="32" tilecount="24" columns="6">
+ <image source="../Art/Tileset props - 16x32.png" width="96" height="128"/>
+ <tile id="10">
+  <animation>
+   <frame tileid="6" duration="200"/>
+   <frame tileid="7" duration="200"/>
+   <frame tileid="8" duration="200"/>
+   <frame tileid="9" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="16">
+  <animation>
+   <frame tileid="12" duration="200"/>
+   <frame tileid="13" duration="200"/>
+   <frame tileid="14" duration="200"/>
+   <frame tileid="15" duration="200"/>
+  </animation>
+ </tile>
+ <tile id="22">
+  <animation>
+   <frame tileid="18" duration="100"/>
+   <frame tileid="19" duration="100"/>
+   <frame tileid="20" duration="100"/>
+   <frame tileid="21" duration="100"/>
+  </animation>
+ </tile>
+</tileset>