+++ assets/Magic Market/Tilesets/Tileset props - 32x16.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="Tileset props - 32x16" tilewidth="32" tileheight="16" tilecount="4" columns="2">
+ <image source="../Art/Tileset props - 32x16.png" width="64" height="32"/>
+</tileset>