+++ assets/Magic Market/Tilesets/Tileset props - 48x16.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.11.0" name="Tileset props - 48x16" tilewidth="48" tileheight="16" tilecount="3" columns="1">
+ <image source="../Art/Tileset props - 48x16.png" width="48" height="48"/>
+</tileset>