+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Take into account the layer visibility information from Tiled and set Bevy Visibility component accordingly
+- Add the 'Magic Market' tileset and example to showcase a more realistic / complex map
 
 ### Bugfixes